requirements.txt
argon2-cffi==20.1.0
cffi==1.14.0
Django==2.2.13
django-cors-headers==3.3.0
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
